refactor(functions): extract swaptoparam
